gh-153569: Split tokenizer lexer into focused files
